requirements.txt
markdown>=3.5
pygments>=2.17
jinja2>=3.1
pyyaml>=6.0
